Clear cyclic references on transferring blocks
Blocks can transfer from its workspace to another in the following steps.
  1. Encode blocks as XML.
  2. Creating a new block on the target workspace by decoding block XML.

All the cyclic reference on the blocks must be removed before the step 1.
Otherwise, variables on a new block created in the step 2 could refer to
variables on the old block because reference relations are also stored in
block XML.

Several tests has failed from changes in commit b2a87abc, but these tests are
successful with this patch.