Add test to show type-expr of isolated reference
block could be out-of-date

It's not a bug. This situation is expected not to happen in the typed-blockly
app.

Reference blocks are created only from workbench in the app. Workbench copies
the same type-exprs with value blocks ans store them to created reference
blocks. (See commit 4b5ee2fe). So the type-exprs would always be up-to-date.